move background grouping functions
